dom/tests/mochitest/pointerlock/file_cursorPosEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_cursorPosEvents.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #parent > div { width:100px; height:100px; }
    #parent { width: 200px; height: 200px; }
  </style>
</head>
<body onload="setup();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>
  <p id="display"></p>
<div id="content">
  <div id="parent">
    <div id="child" style="background-color:Green;"></div>
  </div>
  <pre id="test">
    <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Test will check to make sure that the following mouse events are no
       * longer executed in pointer lock.
       *  - mouseover, mouseout, mouseenter, mouseleave
       */
      SimpleTest.waitForExplicitFinish();
      var moved = 0;
      var pointer = navigator.mozPointer;
      var parent;
      var child;
      var mouseEnter = false;
      var mouseLeave = false;
      var mouseOver = false;
      var mouseOut = false;

      document.addEventListener('mozfullscreenchange', mouseCursorRequired,
        false);

      function mouseCursorRequired() {
        if (document.mozFullScreen && document.mozFullScreenElement === parent){
          pointer.lock(parent, cursorTest);
        } else {
          SimpleTest.finish();
        }
      }

      function cursorTest() {
        if (moved == 0) {
          moved++;
          parent.addEventListener('mousemove', cursorTest, false);
          synthesizeMouseAtCenter(child, { type: 'mousemove' }, window);
        } else if (moved == 1) {
          moved++;
          synthesizeMouseAtCenter(parent, { type: 'mousemove' }, window);
        } else {
          ok(!mouseEnter,
            "mouseenter should not happen in Full Screen and Pointer Lock.");
          ok(!mouseOver,
            "mouseover should not happen in Full Screen and Pointer Lock.");
          ok(!mouseLeave,
            "mouseleave should not happen in Full Screen and Pointer Lock.");
          ok(!mouseOut,
            "mouseout should not happen in Full Screen and Pointer Lock.");
          pointer.unlock();
          
        }
      }

      function setup() {
        parent = document.getElementById('parent');
        child = document.getElementById('child');
        
        document.addEventListener('mozpointerlocklost', function() {
          document.mozCancelFullScreen();
        }, false);

        // Events that only exist because a mouse exists
        child.addEventListener('mouseout', function() {
          mouseOut = true;
        } , false);

        child.addEventListener('mouseleave', function() {
          mouseLeave = true;
        } , false);

        child.addEventListener('mouseover', function() {
          mouseOver = true;
        } , false);

        child.addEventListener('mouseenter', function() {
          mouseEnter = true;
          } , false);

        startTests();
      }

      function startTests() {
        SimpleTest.waitForFocus(function() {
          parent.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
